+++ src/TIClient.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 namespace ATreschilov\TinkoffInvestApiSdk;
 
-use ATreschilov\TinkoffInvestApiSdk\Services\InstrumentsService;
+use ATreschilov\TinkoffInvestApiSdk\Services\InstrumentsServiceDecorator;
 use ATreschilov\TinkoffInvestApiSdk\Services\MarketDataService;
+use ATreschilov\TinkoffInvestApiSdk\Services\MarketDataServiceDecorator;
 use ATreschilov\TinkoffInvestApiSdk\Services\OperationsService;
+use ATreschilov\TinkoffInvestApiSdk\Services\OperationsServiceDecorator;
+use ATreschilov\TinkoffInvestApiSdk\Services\UsersServiceDecorator;
 use ATreschilov\TinkoffInvestApiSdk\Services\UsersService;
 use Grpc\ChannelCredentials;
 use JetBrains\PhpStorm\ArrayShape;
@@ -21,15 +24,24 @@ class TIClient
     private string $token;
     private string $cert;
 
-    private UsersService|null $userService = null;
-    private OperationsService|null $operationsService = null;
-    private InstrumentsService|null $instrumentsService = null;
-    private MarketDataService|null $marketDataService = null;
+    private TIClientOptions $options;
 
-    public function __construct(string $token)
+    private UsersServiceDecorator|null $userService = null;
+    private OperationsServiceDecorator|null $operationsService = null;
+    private InstrumentsServiceDecorator|null $instrumentsService = null;
+    private MarketDataServiceDecorator|null $marketDataService = null;
+
+    /**
+     * @param string $token
+     * @param array{
+     *      isRateLimitRetry: boolean
+     *     } $options
+     */
+    public function __construct(string $token, array $options = [])
     {
         $this->token = $token;
         $this->cert = file_get_contents(self::CERT_FILE);
+        $this->options = new TIClientOptions($options);
     }
 
     #[ArrayShape([
@@ -38,7 +50,7 @@ class TIClient
         'grpc.ssl_target_name_override' => "string",
         'update_metadata' => "\Closure"])
     ]
-    public function getOptions(): array
+    public function getApiConfig(): array
     {
         return [
             'credentials' => ChannelCredentials::createSsl($this->cert),
@@ -56,10 +68,10 @@ class TIClient
         return self::INKOFF_INVEST_API2_HOST;
     }
 
-    public function getUser()
+    public function getUser(): UsersServiceDecorator
     {
         if ($this->userService === null) {
-            $this->userService = new UsersService($this);
+            $this->userService = new UsersServiceDecorator($this);
         }
         return $this->userService;
     }
@@ -67,24 +79,29 @@ class TIClient
     public function getOperations()
     {
         if ($this->operationsService === null) {
-            $this->operationsService = new OperationsService($this);
+            $this->operationsService = new OperationsServiceDecorator($this);
         }
         return $this->operationsService;
     }
 
-    public function getInstruments()
+    public function getInstruments(): InstrumentsServiceDecorator
     {
         if ($this->instrumentsService === null) {
-            $this->instrumentsService = new InstrumentsService($this);
+            $this->instrumentsService = new InstrumentsServiceDecorator($this);
         }
         return $this->instrumentsService;
     }
 
-    public function getMarketData()
+    public function getMarketData(): MarketDataServiceDecorator
     {
         if ($this->marketDataService === null) {
-            $this->marketDataService = new MarketDataService($this);
+            $this->marketDataService = new MarketDataServiceDecorator($this);
         }
         return $this->marketDataService;
+    }
+
+    public function getOptions(): TIClientOptions
+    {
+        return $this->options;
     }
 }